Added abstraction for callee to take advantage of massage_dispatch.
Two different types is used for caller and callee to handle the same message. Due to different semantics in who owns the buffer. Caller side does not own the buffer (lends it from the buffer pool, and the user is still responsible for the buffer). For the callee side, the message owns the buffer (since it's has been deserialized from the queue) and then moves it to the pool.