Added Scope - to enable scope within a casual-make-file
This is brilliant, IMHO. It enables the user to scope a group of
casual-make-directives with a new "context"

Under the hood the scope is generated in a isolated makefile, and
invoked from the parent scope.

The use-case I aimed to solve is concurrency. That is, to enable the
user to group directives in a casual-make-file where some is processed
in parallel, and others not.

The only directive that makes real sense to put in a scope is
Build(...), that I right now can think of. But, technically the user can
scope with a new set of 'library-paths' and such.

The user can have as many scopes he wants. The scopes can be nested in
arbitary depth

Example:

Build( "common/makefile.cmk")
Build( "xatmi/makefile.cmk")
Build( "buffer/makefile.cmk")
Build( "serviceframework/makefile.cmk")
Build( "configuration/makefile.cmk")

with Scope():
    Parallel( True)

    Build( "tools/makefile.cmk")
    Build( "broker/makefile.cmk")
    Build( "monitor/makefile.cmk")
    Build( "transaction/makefile.cmk")
    Build( "queue/makefile.cmk")

Build( "test/makefile.cmk")