casual-make Compile - objectfile is now optional
If objectfile-name is not provied, one will be created autmatic, in a
platform specific way.

Since Compile returns the 'object-file-target' and this return is the
thing one links, I think the samantic is still good enough to understand
what is going on in reality. Either way, I think it's easier to write
and mantain the casual-make-files this way. And, one can still
explicitly specify the object-file-name if one wants...