RM-configurations can be greater than linked RM:s
Before, the result of all RM-configuration for groups that the server is
member of, had to be an actual linked RM.

Now we base the search for configuration on linked RM:s.
This makes it really casual to configure a domain if one wants to: just
make one group with all RM configurations and make all servers member of
that group. (this does not work yet though, since groups can only have
0..1 resources -> we need to change this)