Admin - terminal-archive: generic terminal output
Got tired of writing explicit code to output to terminal. A lot of work
to align, color, headers and so on.

Wrote a "generic" archive that take cares of:
- column width
- header (on/off)
- color (on/off)
- alignment (right/left)

Takes a vector of Directive:s where one can give explicit directive for
0..N columns, denoted by attribute (role) name. That is, color and
alignment.
Also, two booleans that dictate header on/off, and colors on/off