WIP correlate calls with transactions
So we know if we have to receive pending calls during commit. Likewise,
we know which calls can be discarded by the user.

Also, the correlation between the call-descriptor and pending reply is
now based on the actual correlation in the message. Much more safe and
strict, and paramount to be able to discard messages.

Got rid of common::call::Context own pending-reply-cache, we now rely on
the ipc-abstraction's cache. Easier and more robust, consumes less
memory and perhaps a bit faster.