Fixed 'signal-safe' signal handling.
This should fix the "1/10000" deadlocks that occurred when a signal was
received while we already had taken a lock to the 'signal-structure'. We
now only use std::atomic which is standard conformant when dealing with
signal-handlers. I was an idiot to think that a semaphore and a lock
could ever work...

Some minor refactoring in exception, change 'severity' to 'category' so
it correlates to log-category (which was the purpose in the first place,
before we change 'level' to 'category' for logging)