feature: casual-forward-cache forwards 'send & forget'
Part of casual-broker, to enable forwards to services from the same
instance.

Could be extended to forward any type of service, hence act as a cache.
Not sure if we want this though.

Also: improved mockup-ipc - we now have input() and output() that denote
the 'input queue' and the 'output queue', so it's easier to reason about
the semantics when setting up 'chains'.