optimizing casual-queue
As I understand sqlite, we can't use any of the aggregates min, max,
avg, sum
without triggering a table-scan. sqlite only uses index for min, max if
there is no where-clause. avg and sum of course has to calculate an
average or sum, hence table-scan.

Using triggers to keep track of number of messages and the total size of
all messages in a given queue, so we at least can give the user some
info about the size of a given queue.

If sqlite change their implementation maybe we can revisit this then.