gateway #9 Improved signal semantics - rudimentary gateway
signal::handle now takes the current signal mask into account if a
signal is to be handled. This makes the semantics much more easy to
reason about. If you block some signals in a thread you'll never handle
these until the block changes/goes out of scope.

Rudimentary gateway functionality.
unittest start an outbound that connect to the local domain. Works great
and it should work to other domains.