wip doman-broker split, #18 almost done, some work on gateway
Most of the work is cleaning up mockup-abstraction. It's a lot better
now. One can understand the unittests a lot easier with these
abstractions. Removed the old ones, and had to refactor a bunch of
unittests...

Some issues with json-archive unittests, but I need to push this to get
a backup.