interdomain transactions #9 - normalized resource messages
Resources does no longer 'know' if it is the local domain or a remote
domain who has triggered a commit.

If it's a remote domain that instigated the request we keep track of the
RM-id (in the message) so we later can reply with the same RM-id, hence
enable the remote domain TM to correlate the reply.