Some improvements on unittest that sometimes fails.
On OS X there seams to be a glitch when you spawn a lot of processes and
directly send (terminate) signal: Sometimes the signal for the last
spawned process is "lost", that is, the invocation of 'kill' report OK,
but the process does not receive the signal. Don't know why, but it
seams like the process is in some limbo state (not really started) and
somehow the OS "drops" the signal. At least that is the symptom.

I change the spawned processes to the standard "sleep" for these tests,
which seams to work better. Don't know why, but it could be that it's
starts "faster", hence it's less chance for the "glitch"

It's unlikely that this would ever be a problem when running real stuff
in casual, but it's still irritating and if it can happen, it's only a
matter of time until it does happen..