Attempt to fix signal glitches and other stuff to get rid of cores
Simplified the signal handling a lot. Got rid of some hack that probably
was the cause for cores.

When we spawn a process we now sleep for 200us to minimize the problem
of the spawned process not receiving signals during some small time.
This is not a solution, but if it works good enough for now we can keep
going, and try to solve it later on.