Fixed bootorder and block of signal propagation to children
Got rid of hack when spanning child, seems to be working stable.

In process::spawn we now make sure that child is in another process group,
hence does not get signals that parent receives. Before it was impossible to
order shutdown since pretty much every process gets the terminate signal and
starts to shutdown by it self.

Got rid of signal blocking during spawn, since (as I understand it) it's guarantied
that signals received during posix_spawn will be retried after the spawn is
completed. It seems to work at least on my machine.

We now have a few groups to make sure we have correct dependencies between
casual stuff, and users dependency to casual.