bash-completion to simplify issue #107
This is the first version and somewhat rudimentary. Right now it doesn't
really gonna work :) Since the option `casual-bash-completion` is going
to be found for `casual-admin` and trigger there. I need to either
implement some _grouping_ or to have a special `directive` for spawning
_sub/child admin clients_ so I can forward the `casual-bash-completion`

see common/documentation/arguments.development.md how to use it.