issue 131: bash completion not working on OS X
compopt is not available in bash 3.2, hence it is not
possible to change options "on the fly" (as far as I've been
able to tell).

these changes makes the completion compatible with bash 3.2,
however, when auto-completing configurations files for --boot, a
space will be added for completed directories.

some clean up has been done as well.

tested on centos 7, with bash 4.2.46 and 3.2.57.