make sure queue-forward has correct dependencies in unittest
if _queue-forward_ and _queue-manager_ are associated to the same
group there is no guaranties whom shutdowns first, which can
lead to "dead-lock", since _queue-forward_ tries to "disconnect" to
"dequeue-event" during shutdown. If _queue-forward_ happens to send
the "forget-request" and _queue-manager_ happens to shut down at the
"same time", _queue-forward_ will never get the "forget-reply"

Made sure _queue-forward_ is member of a `group` that is dependent
to the `group` that _queue-manager_ is member of.

We might need to make this more robust.