argument - separated assignment and invocation
first assignment pass, then invocation pass
group can register a callback for 'invoked' and can act on invocation
(instead of just a "holder")

Also, a few CLI changes regarding "buffer piping" so we can pipe several
buffers. Can be useful "in production", and is very useful for demos and
such.