compatibility suse g++ (version?) constexpr static member variable
As I see it we don't use the constexpr static member variable in an
odr-used context, so we should not need to define it at namespace level.

C++17 takes care of this with inline constexpr variables

Changed it to inline  constexpr static member function instead. Until
we use c++17 this is probably the best way.