bug: cobol binding used a flag that is not valid for tp(a)call
TPTRAN, which is defined as flag in xatmi.h, but is not valid for
tp(a)call. I think the reason for this bug is that it was/is defined.

There is no reference to TPTRAN in the C-api-section of the xatmi specification,
hence I think we should remove the definition in xatmi.h.