range::adapter to help iterate over sub-ranges of ranges
We have a few use-cases where this get rid of raw loops, and make
the code more precise (well it's somewhat subjective...)

It started as _code masturbation_, but turned out better than expected,
so we keep it. I expect there are more use-cases we could use it to
improve readability.

Not sure about the name though... range::adapter::next might be better?
I'm pretty sure the range::adapter part is correct, but we might need a
more descriptive name.