service parameter logging working as it should
This haven't worked for a while. One might argue that it's not an
important feature, but it's pretty cool and demonstrates the separation
between data structures and format/serialization.

The implementation for the actual feature is simple, but the _flush_
semantics on archives/protocols was broken. Had to refactor a whole lot.

The new semantics are explicit, user needs to explicitly say "consume
the serialized information to this destination", and the archive needs
only to have the internal state that matters for it.

Most of our archives has to keep an implementation defined structure
that is bound to the 3pp we use. So, before we had the "illusion" that
the archive actually wrote to the wanted destination "live", but in
reality it just kept a reference to the destination and did the write on
_flush_.

_flush_ was called in the dtor, and some time explicitly if it "was
needed". This was extremely hard to reason about. Some archives didn't
work as expected if flush was called multiple times. For example: xml
what does it mean to _flush_ "part of a document", and then continue to
serialize to the document? The result is either two separate documents,
or two documents where the second also contains the first. One might
expect the end result to be one document with no duplications. Hence,
even the unexpected result is ambiguous and hard to reason about.

I choosed the name `consume` - not totally sure if this is the best name
for this. It's rather the other way around. If we find a better name we