bug domain task did recursive callbacks
made sure we only invoke task callbacks from within a message handler,
hence we get a strict boundary between tasks and other stuff.

detail:
The only place we did invoke callbacks was when we spawned a process.
Instead of dispatch the event directly, we push it to inbound and have a
message handler take care of the dispatch later on.