domain - added group restart
if no groups are provided 'all' aliases are restarted.

filtered out all `casual` internal servers and executables, hence it is
not possible to restart `casual` stuff. For now... It is not given what
to do if stuff is mid-flight, so we just disable the possibility to
restart `casual` stuff.

also:
* added _deprecated_ argument option semantics
* added _deprecated_ service _context_ -> clear what is deprecated or
not.