domain runtime configuration core functionality
Refactoring so that we have 3 "models" that represent configuration and
state

* user model - focus on what is good for the user
* model - the internal model - what is good for the system
* state - the actual state in each 'manager'

Only `domain-manager` "knows" of the user model, the rest of the system
only knows of the model and its local state

We transform in the following way:
user <-> model <-> state

We use the difference between current model and wanted model to deduce
what to do, to achieve the wanted state.

Note: nothing is actually done to use this functionality yet. The plan
is to introduce runtime configuration one configuration "object" at a
time. It is not trivial to get from one state to another in some
cases, especially in an asynchronous way...