transaction bug - did not remove 'consumed' transaction
Over time this consumes a lot of memory and a lot of work (find
transactions) is done in vain.

This is really old code, and it's strange that we haven't noticed this
before.

The transaction semantics was correct though, and this is probably why
we haven't found the bug before.