event - added event that signals when a discoverable is 'on-line'
This enables blocking/waiting lookups to issue a discover when a new
discoverable is on-line. Hence, it's possible to wait for _resources_ in
other domains.

This only (?) effects queue.forward waiting lookups right now. Or
rather, we only use wait-lookup from queue.forward.

Example: queue service forward is configured to call service `x`, that
is not available in local domain. Some time later local domain get an
outbound connections to a remote domain, this will trigger a discovery
for the pending lookups to the new discoverable, and if the remote
domain has service 'x' the forward will start working.