cli - fixed call to keep correlation id for each logical message
`call` sent part of the logical message, and got a blocking issue, when
the message was resent a new correlation id was created, and the other
side (rightfully) treated this as a separate message. Hence, 'the other
side' inbound cache grew accordingly. Under heavy load against
gateway-outbound-group eventually most of the work that was done was to
'traverse the cache' with dead partial messages.