gateway - improved connect to be more robust
We pretty much accept all addresses and just try to reconnect regardless
of the type of error, if any.

It's very hard to separate 'fatal' errors from transient errors. So, in
the spirit of `casual` we just trust the user and the setup for the
network topology and try to connect every 3:d second, regardless. It's
probably more intuitive for the user also, to see all configured
connections with `casual gateway --list-connections`