gateway - improved tcp connection phase robustness
When inbound/outbound-groups connects with a peer, it now spawn a process
that takes care of the logical connection. Hence, this process can care
and make all kinds of mess and it will not affect the other established
connections.

Also:

* made sure we use a copy of all read/write ready descriptor
ranges

Before: When we're in the middle of iterating over ready reads we might
get a message that a new connection is done with the logical connection
phase, and we add the socket (and other stuff) to 'running' connection,
AND add the descriptor to the read state, which might realloc the
container, hence, we (possible) continue to iterate over memory which is
_dealloc_

* gateway - only remove the 'transaction mapping' when all branches are
removed

We removed the mapping and all the branches when the first branch where
removed. If there where more than one branch to external domains we'll
log an error when the next commit/rollback reply arrives.

If the commit/rollback requests for all branches to on gtrid arrives to
outbound with enough time between then that a reply have time to get
consumed, we'll fail to correlate which connection we should send the
commit/rollback to. Highly unlikely, but it could happen, and this would
not be good...

* changelog

Approved-by: Hans  Bergkvist