gateway - false error during tcp consume of logical (large) message
When there's a `pselect` dispatch we start consume the message within the tcp handler.

If we don't succeed to consume a _complete_ message we classify this as an error.
Which is a false error, since the rest of the message could arrive later. The
bigger the messages the more likely it is to only get parts of messages, of course.