configuration - service manager runtime configuration update
Works somewhat...

The implementation in service::manager::configuration is **way** to
complicated, and I don't think it's actually works in all cases.

We need to make the state and/or the update semantics a lot simpler.
I'm convinced that this is possible, but how is not given, and need
some thoughts...