queue - forward to remote queues did not work correctly
The discovery worked correctly, but the remote queues does not have
a queue::id.
Added a general remote::queue::id to associate with remote queues, so
the rest of the code works as before.

Also, made sure we only reply with local queues for a discovery, unless
it's a "forward-discovery".