xatmi - fix leak of descriptors for conversational services.
The descriptor allocated (reserved) on tpconnect was not "unreserved" when
a conversational service terminated. This is a resource leak. Showed up as
a message in the casual log on server shutdown. Also as growning descriptor
numbers passed to service when called.

Also Small improvments of tests for conversational:
* Verify that descriptors are reused.
* Renamed service casual/example/conversation2 to casual/example/conversation_recv_send to indicate that it does
tprecv and tpsend.

Approved-by: Fredrik Eriksson