conversation - generate events in tpsend when partner calls tpreturn or tpdiscon.
Added tpsend detection of subordinate having done tpreturn. Conceptually
this is a service side abort of the conversation.

Handle the case of a tpsend in subordinate detecting that
initiator has disconnectd the conversation. Conceptually an initiator
side abort of a conversation.

A number of unit tests added where subordinate does "odd" things. Test
service is told what to do by the unit test code.