log - made sure we only use our own stream::write and got rid of global ostream stream operator
It was way to fragile and probably UB - different behaviours on
different platforms and versions of compilers.

When streaming on error paths with gtest it's probably easiest to use
CASUAL_NAMED_VALUE( <value to stream>) to hook in to `stream::write`.
Also added a `compose` function in the `unittest::Trace` type for
convenience. "All" unittest has a `unittest::Trace` instance, and can
just call `trace.compose( ...)` and `stream::write` will be used.