unittest - fixed gateway unittest - some unittest refactoring
Reworked gateway test_maanger.cpp unittests since they relied on be able
to connect to the same domain, which does not work any more - we
actively _terminate_ the discovery if the destination is the same as the
domain that does the discovery.

The objective is to have all "managers" provide unittest friendly stuff
for the things that each manager manage exposed for others to make it
easy to construct unittests.

I did a few that was needed, but there are a few left...