queue - forward react if queue group "dies"
Before there could be "race condition" if a dequeue message is sent from
a forward while the queue-group is shutting down. That is, the forward
first send _dequeue_ and then _dequeue-forget-request_ (but queue-group
is not consuming these messages - shutting down), then the queue-group
exit, and queue-forward will wait forever on _dequeue-forget-reply_.

With this fix: queue-forward listen on _process-exit_ events, and if a
dead process is a queue-group that the queue-forward is waiting for a
_reply_ from, we just "remove" that forward, hence breaking the
deadlock.