Rebased to release/1.5 on august 15 2022.
Added tpsend detection of subordinate having done tpreturn.

A unittest where subordinate does a tpreturn when not in control of
the conversation added.

This change also handle the case of tpsend in subordinate detecting that
initiator terminates conversation with tpdiscon(), This has not been
verified.

Some rerstructuring of code in ewample server, still WIP