WIP! Code formatting, More test cases.
reformatting and some code clean up.

Added an action to example server to ask it to do
  tpreturn(TPFAIL, ..., data, ...)

Changed "actions" in conversational "test" service to only
look at last message received instead of collected data
from invocation data and receives.

Added a test to test_converation.cpp to test the case of
service doing tpreturn(TPFAIL, 0, data,...) when not in control. It
is currently DISABLED_ as test fails.