transaction - fix distributed transaction bug in complex topology
More or less a total rewrite of transaction-manager. We now use
coordination for _fan out_ multiple request, and when we got all replies
we do a 'continuation' based on the specific context of the original
request. Hence, we know exactly what to do, and this with less state and
less code in general, which is really good.