Be conformant with flags to xa_end
Make sure we always uses exactly on of [ flag::xa::Flag::success,
flag::xa::Flag::suspend, flag::xa::Flag::fail] on xa_end.

Right now we never set _fail_ since this would make it impossible to
interact with the resource after. All units of work will be rolled back,
but the user might want to interact with resources for reads after a
service has return with _fail_. Well, it's possible a quality of
implementation issue for the resource, but at least db2 makes it impossible...

fixes #28