ipc - added multiplex send abstraction
Did not change the majority of current semantics, tried to add it "on
the side"
If this works well, we probably will only use the multiplex send.

One drawback is that we'll be told by _pselect_ that there are file
descriptors ready to be written to, but the size we have to write/send
is larger than the space the file descriptor has (we send transport
messages of a known size, to eliminate interleaving writes). Hence we
will not block in _pselect_, and try to write in vain, until the other
side consumes messages on their side.
The upside is that it will probably be very few "busy sends", at least
when we use this "everywhere". And the most important thing is that
we'll be in the message pump all the time, and process other messages,
hence move forward, and get things done.

It is possible that _pselect_ will block if/when we're using FIFO/pipes
as IPC mechanism. I think the total size of the "buffer" is a multiple
of PIPE_BUF, but there is still the "last transport" of a complete
message, that is most likely less than PIPE_BUF, so we probably get the
same problem still, but less frequent.