transaction - fix external resource prepare/commit request
Before:
We had a generic function for prepare and commit, and this does not work
since we never can do _one-phase-optimization_ for a prepare request.

After:
We now have differnt semantics between prepare and commit. We only do
a _one-phase-commit_ in the externa commit handler, iff there are one
resource involved.