gateway - fix outbound to allways remove transaction state
Before:
We did not remove transaction state if the resource prepare was
read-only. We only looked at commit/rollback

After:
We do look att resource-prepare-reply in outbound, and if the message
got _read-only_ we know that the transaction (branch) is done, hence we
remove it from the "mappgin state".
Result: We don't "leak" state, and the state for "transaction mapping"
will not grow over time, hence performance will not decrease over time.