gateway - fix inbound to stop consume from a removed connection
Before:
When inbound get domain::disconnect::Reply (from the previous Request)
AND there's no pending request/reply in-flight, the connection is
removed/destructed. The _select dispatch_ still might try to consume more
messages from the _destructed_ connection, which of course led to
errors.

After:
When the handler removes/destruct the connection, it also notifies the
_select dispatch_ to stop consume from the connection (via returning
false).