configuration - fix configuration normalize
Before:
We produced the internal model instead of _users model_, which is not
something the user should see/know about. The internal model is of
course not compatible with the _user model_, hence the output of
`--normalize` could not be piped to other cli stuff.

After:
We produce _user model_ and added tests for normalize