environment - fix environment::set for std::filesystem::path
Before: the std::filesystem::path got "serialized" via ostream when
setting a environment variable. Hence, the path got "quoted". When
getting a path, it's just a regular string we construct
std::filesystem::path with. This result in some very strange actual
paths.

I don't think it's likely to happen in "production", but our unittest
creates a bunch of weird directories.