conversation - fix a bug where tprecv and tpsend failed to set the global tpurcode.
This is the user return code specified when a subordinate service in a
conversation calls tpreturn().

Also some work on unit test cases. More test cases to add, but this change is likely to work.

One error remains. When subordinate calls tpreturn when not in control of the
conversation the xatmi spec say that the initator should get a TPEV_SVCERR event.
It currently gets a TPEV_SVCFAIL event. The unit test for this scenario is "DISABLED".