queue - fix local lookup to also search external known queues
Before: From a local lookup we only searched internal queues to the
domain, hence a discovery was always triggered for external queues.
When the discovery replied, a more general search was done and the queue
was found, and it sorta worked.

unittests for this is in progress, but not done yet.