gateway - immediately send error reply on failed resource request
Before:
We took no immediate action when failing to correlate a descriptor
with a connection when sending resource requests.
This would occasionally lead to situations where casual-queue-forward
would hang waiting on a commit/rollback reply.

Now:
We confirm that the message was successfully sent and immediately
invoke the error-callback if not.

Fixes #98