queue - add browse-peek to queue api
Add _browsing_ functionality to the queue api. User provided callback gets
invoked with a message until the queue is empty or the callback returns
false.

Corresponding C-api added as well.

We could expand on this in the future and have the same semantics for
browse-dequeue. It might be a better api for the user.