discovery: add state and metric information to discovery
This patch adds a state service that exposes some of the state the
discovery has.

The most interesting state is counts of messages sent and received,
which by it self is probably a good metric to have, and we can deduce
other stuff that might be more of interest to the user.

This is probably something we want to generalize to all _managers_.

resolves essue: #110