queue: handle enqueue errors correctly
There are no error state in the dequeue::Reply, and the only way to
indicate error is to propagate a 'nil' message id (this is not
sufficient, and need to change to 1.7).

Before: We did not check this _error indication_ both in the API and in
queue-forward. Which could lead to "lost messages".

After:
* Forward: We check this _error indication_ and assumes
   `queue:no_queue`. Invalidate the queues, and "restart" the forward
   "flow"
* API: We check this _error indication_ and raise `queue:no_queue` if
  'nil'