tcp: fix partial header offset to be a multiple of char, not header type
This patch fixes a bug that when sending partial header, the offset and
size was based on the size of the header type, instead of char.

Also: posix does not define any block or no_block flags to sendmsg and
recv. Made sure we set and unset O_NONBLOCK on the filedescriptor for
blocking and non blocking.