queue: fix queue-forward process exit clean up
When a process exits, and the process is a group that queue-forward has
pending (de|en)queue messages, queue-forward emulates "empty" replies to
it's own inbound, and let the normal flow end naturally.

Before: There was direct cleanup in the process-exit-handler for the
pending messages and such, which made it "impossible" to act correctly
in the normal flow later, when dealing with the emulated replies.

This patch removes the direct cleanup.