common: fix UB for transcode::hex encode/decode
Before:
transcode::hex encode/decode reinterpret_cast the source and/or target
to `char*`. This was done by dereference the _iterator_ and then take
the address which was casted. Since we use iterator semantics we had
_begin_ and _end_, and dereference _end_ is UB (end is one past the
actual container).

Now:
Moved the actual hex stuff to the header and we don't have to cast to
some common type.

Note:
This doesn't seems to have caused any trouble but as we know it's UB,
we'll of course fix it.