buffer: fix field doesn't track 'service buffer' during auto realloc
The buffer that a service is invoked with is 'special' according to the
xatmi spec. casual needs to keep track of it and dealloc when the
service is done.

buffer::field auto realloc the buffer during add and such, without
pool::Holder knowing about it -> pool::Holder failed to correlate the
_service inbound buffer_ during service finalize -> error and
potentially memory leaks.

This patch adds ability for buffers to help pool::Holder to keep track
of the _service inbound buffer_. And make sure buffer::field does it.

Resolves #179