queue - prevent forward causing service-manager coredump
Before: Occasionally, the forward would request a discard of a
service lookup but still call the service, depending on the
order of the replies. This would trigger a core in the
service-manager when the service ACK:ed the call.

Now: We handle pending service discard requests separately so
that no service call occurs irrespective of the order of replies.

Fixes #182