service: service timeout could render zombie transactions
Before: When a service trigger a timeout by service-manager, an error
reply is sent from SM. caller will then rollback the transaction.
The service is till running (if linger) and could trigger resource
involvement with the transaction (again).This new transaction
(it's new to TM since the TM has forgotten the recently rollbacked
transaction) will never be handled, and will most likely be "marked
as in-doubt" -> cause a lot of unwanted problems.

Now: When service manager receives an "ACK" from the timouted service it
sends a rollback request to TM -> If there are resources that has been
involved the rollback will be issued, exactly like a normal rollback.

Note: it feels a bit ugly to let SM actually order a rollback, it might
be better if it "just notify TM", and TM decides what to do. Like a
"post timeout cleanup rollback" message. The current solution is less
intrusive though.

Resolves #191