ipc: improve detection of removed ipc devices (files)
When a process exit (could be a core/signal) DM tries to remove the
associated ipc socket file. This to enable processes to detect that an
ipc device is gone, and handle potential consequences.

Before: DM did not remove the ipc socket file properly (or at all?)

Now: DM removes the ipc socket file, and others can detect this.

resolves #194