gateway - handle routing problems (#197)
Before:
When discovery requests still beeing in-flight and only some of the destinations had replied and a connection lost was received on a connection that already had replied, that reply still were wrongly considerd as a valid provider of the requested service. That caused faulty entries in the service lookup list.

After:
When a connection lost is received all replies in an ongoing discovery from that connection/descriptor are considered failed.