gateway: outbound order of connection for a service gets too "orderly"
Before: The order of connections a given service has (the remote domains
that expose the service) is based on the order which the connections
connect. In some topologies this could give a bad load balancing.

Now: When we add add a connection to a service we insert it a random
position. round-robin is used as before after that.

Resolves #201