discovery - handle prospect services (#198)
Before:
In-flight discovery request was not considered upon rediscovery requests and explore resulting in missing services ( and queues) routing to new connections.

After:
Every in-flight discovery request registers it's resources to a prospect list which is considered when doing rediscover and exlpore.
After the discovery request is done all prospect claims are removed from the prospect list made by that correlation.