discovery: aggregate external discovery requests for performance
This patch tries to aggregate discovery::Request based on load. The load
is measured as _current discovery request in-flight_.

We've got a _load-window_ that defines how many in-flights to trigger
the "next level". Every "level" adds a _timeout_ where every external
discovery request are accumulated within.

When the _timeout_ kicks in, the aggregated request is sent, and later
when the accumulated reply arrives it is _fan out_ to corresponding
replies to match the actual original requests (correlation id and such).

"level" semantics:

The first "level" is `0`, hence has no accumulation and no timeout ->
every request is sent "as is".

"level" 1 has a timeout with accumulation until send of 1 x _timeout
value_.

"level" 2 -> 2 x _timeout value_.

... and so on.

To reach the "next level" the load has to increase with the
_load-window_.

Example: Lets say _load-window_ is 2, and _timeout value_ is 10ms.
If we have a load of 3, a timeout is set to 10ms, and we start
accumulating request. When the timeout kicks in, we send the request,
and invalidate the timeout.
A new request is about to be "sent", we check the load and decide if
we should accumulate or not. We could have got replies during this time,
and the load has gone down.

Note: To reach "level 2", there need to be a load of 5 -> including at
least 2 accumulated requests -> at least 2 x 10ms without "any" replies.

This "algorithm" should self tune against heavy loads, at lest to other
domains over network.

Temporary "config", will probably be renamed or removed in the future:
* CASUAL_DISCOVERY_ACCUMULATE_REQUESTS = <load window default 20>
* CASUAL_DISCOVERY_ACCUMULATE_TIMEOUT = <timeout default 4ms>