discovery: replies should only contain a subset of the requested
Before: when we fan out the replies for accumulated discovery replies
from other domains we do not filter base on the requested services and
queues. Hence, the replies to caller could contain services and queues
that was not asked for. Which lead to domains knows about
services/queues that there hasn't been a need for.

Now: The accumulated reply content is intersected with what the actual
caller has requested -> a subset of the requested is replied to caller.