http: fix nginx file/buffer handling on requets
Before we didn't handle multiple buffer/files correctly (well at least
nginx didn't like the way we did it).

Now: we till nginx that we want the request in one buffer (or one file),
which simplifies the whole request buffer handling.

Still, it is not given how nginx handles this, where our responsibility
starts and ends. At least to me, it's not clear if this works in all
situations.

resolves #242