common: add message type count metric for send/receive
Added functionality to keep track of number of messages sent and
received for each message type.

Every time communication::device::Inbound/Outbound receives or sends a
complete-message the counter for that message type is incremented.
* if sent, sent counter is incremented
* if received, received counter is incremented

* Added request and reply message  to extract the aggregated list of
sent/received messages.
* Added a default handler for the Request to reply with the global state
  of this metric.

This makes it possible to ask any casual process for their message
count.

This can be very useful in unittest, but also to help us to produce good
admin cli info -> We can deduce stuff in-flight. Number of specific
events, and so on.

We probably want to expose this as generic cli in the `internal`
_argument group_.