log: do not throw if we can't open logfile -> error on stderr
This patch makes sure that if we can't open/reopen/relocate casual log
file we write an error to stderr. Before an exception where thrown.

We still don't check if an actual log write is successful or not. If,
for example, the logfile gets "broken", we silently drop log entries.

Resolves #250