cli: fixed and simplified the transaction cli
Before: The whole transaction cli did not work (at all?) to begin with.
It was to complicated, and tried to do some "compound" transaction
juggling.

Now: Much simplier. You start a transaction with --begin, and you need
to end the transaction with --commit or --rollback. The cli stuff
in between are associated with the transaction.

@note This is a large commit, since I did not find a way to split it,
and still compile.

Resolves #272